Refactor: 개행 및 상수 정리
