+++ src/constants/chat.ts
@@ -1,0 +1,6 @@
+export const SUBSCRIPTION_DESTINATION = "/sub/public";
+export const PUBLISH_DESTINATION = Object.freeze({
+  zzal: "/pub/image",
+  greet: "/pub/hello",
+});
+export const CHANNEL_ID = "public";
